refactor: move services to modules
