+++ src/controllers/webhook/ai-hub/erp/update-lead.js
@@ -1,5 +1,5 @@
 import { HTTPException } from "hono/http-exception";
-import LeadService from "../../../../services/erp/lead";
+import LeadService from "../../../../services/erp/crm/lead/lead";
 
 export default class AIHubERPUpdateLeadController {
   static async create(ctx) {
