+++ src/services/erp/contacts/index.js
@@ -1,0 +1,7 @@
+import AddressService from "./address/address";
+import ContactService from "./contact/contact";
+
+export default {
+    AddressService: AddressService,
+    ContactService: ContactService
+};
