+++ src/services/erp/crm/index.js
@@ -1,0 +1,5 @@
+import LeadService from "./lead/lead";
+
+export default {
+    LeadService: LeadService
+};
