+++ src/services/erp/selling/customer/customer.js
@@ -1,4 +1,4 @@
-import FrappeClient from "../../frappe/frappe-client";
+import FrappeClient from "../../../../frappe/frappe-client";
 
 export default class CustomerService {
   constructor(env) {
