+++ src/services/erp/selling/index.js
@@ -1,0 +1,7 @@
+import OrderService from "./sales-order/sales-order";
+import CustomerService from "./customer/customer";
+
+export default {
+    OrderService: OrderService,
+    CustomerService: CustomerService
+};
